requirements.txt
flask==2.3.3
tensorflow==2.18.0  # Adjust based on the version you're using
Pillow==9.4.0
numpy==1.24.3
gunicorn==20.1.0


